Enhanced readability of parsing rules
The `parse_statement` function would consume the current token before
calling the appropriate parsing function for each statement rule. This
lead to said called functions not containing the entire rule, but only
the part after the first token. For readabilities sake, I changed it
such that `parse_statement` no longer consumes the starting token of
the rules and handle said token in the rule-appropriate function
instead. Note that this allows rules to be read from top to bottom
without requiring any knowledge about how it was called. On the flip
side this slightly decreases performance as the starting token of a
statement will now be checked twice, rather than once.