Moved the snapshotting of marks into codegen
Previously the snapshotting logic was embedded in the "CFG" during lowering.
This dilluted the CFG with implementation details, since instead of the snapshotting being a feature of mark nodes it was baked into their neighborhood in the CFG...
Now the associated combination of sink and source CTEs is implicitly generated during code generation.