Dropped multi nodes
I use the term "multi nodes" here to describe the way Assignments and Emits nodes functioned, i.e., that they could contain more than one assignment or emit respectively.
This is only usfull for optimization purposes, but we do not need to optimize since this type of optimization is well within the capabilities of a DBMSs query optimizer.